Change alerts notifications handler to other controller file
